refactor: improve dynamic part of the app
